pyproject.toml
[tool.poetry]
name = "self-study"
version = "0.1.0"
description = ""
authors = ["Sergius Art <xcxserj@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
ipython = "^8.21.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^69.1.0"
django-filter = "^23.5"
coverage = "^7.4.3"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
